create VideoDetails component, fix description line breaks
