+++ packages/atlas/src/views/viewer/VideoView/VideoDetails.styles.ts
@@ -1,0 +1,79 @@
+import styled from '@emotion/styled'
+import { Link } from 'react-router-dom'
+
+import { Text } from '@/components/Text'
+import { Button } from '@/components/_buttons/Button'
+import { SkeletonLoader } from '@/components/_loaders/SkeletonLoader'
+import { cVar, media, sizes } from '@/styles'
+
+export const DetailsWrapper = styled.div`
+  ${media.md} {
+    padding-left: ${sizes(14)};
+  }
+`
+
+export const DescriptionContainer = styled.div`
+  margin-top: ${sizes(6)};
+  margin-bottom: ${sizes(8)};
+
+  ${media.md} {
+    margin: ${sizes(8)} 0;
+  }
+`
+
+export const DescriptionTitle = styled(Text)`
+  margin-bottom: ${sizes(2)};
+`
+
+export const DescriptionCopy = styled(Text)`
+  display: block;
+  word-break: break-word;
+`
+
+export const Category = styled(Link)`
+  display: flex;
+  align-items: center;
+  text-decoration: none;
+
+  svg {
+    margin-right: ${sizes(2)};
+  }
+`
+
+export const LicenceCategoryWrapper = styled.div<{ detailsExpanded: boolean }>`
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-template-rows: 1fr;
+  gap: ${sizes(6)};
+  visibility: ${({ detailsExpanded }) => (detailsExpanded ? 'visible' : 'hidden')};
+  max-height: ${({ detailsExpanded }) => (detailsExpanded ? 'auto' : '0')};
+  overflow: ${({ detailsExpanded }) => (detailsExpanded ? 'unset' : 'hidden')};
+  opacity: ${({ detailsExpanded }) => (detailsExpanded ? 1 : 0)};
+  transition: opacity 150ms ease-out;
+`
+
+export const LicenseCustomText = styled(Text)`
+  margin-top: ${sizes(2)};
+`
+
+export const ExpandButton = styled(Button)`
+  display: block;
+  margin-top: ${sizes(2)};
+`
+
+export const CategoryWrapper = styled.div`
+  text-align: left;
+
+  path {
+    fill: ${cVar('colorText')};
+  }
+`
+
+export const DescriptionLink = styled(Button)`
+  word-break: break-all;
+`
+
+export const DescriptionSkeletonLoader = styled(SkeletonLoader)`
+  height: 28px;
+  margin: ${sizes(4)} 0 0;
+`
