+++ packages/atlas/src/views/viewer/VideoView/VideoView.styles.ts
@@ -1,6 +1,5 @@
 import { css } from '@emotion/react'
 import styled from '@emotion/styled'
-import { Link } from 'react-router-dom'
 
 import { GridItem, LayoutGrid } from '@/components/LayoutGrid'
 import { LimitedWidthContainer } from '@/components/LimitedWidthContainer'
@@ -83,11 +82,6 @@ export const TitleContainer = styled.div`
   border-bottom: 1px solid ${cVar('colorCoreNeutral700')};
 `
 
-export const DescriptionSkeletonLoader = styled(SkeletonLoader)`
-  height: 28px;
-  margin: ${sizes(4)} 0 0;
-`
-
 export const Meta = styled(Text)`
   display: block;
   margin-top: ${sizes(2)};
@@ -109,67 +103,6 @@ export const ChannelContainer = styled.div`
   }
 `
 
-export const DetailsWrapper = styled.div`
-  ${media.md} {
-    padding-left: ${sizes(14)};
-  }
-`
-
-export const DescriptionContainer = styled.div`
-  margin-top: ${sizes(6)};
-  margin-bottom: ${sizes(8)};
-
-  ${media.md} {
-    margin: ${sizes(8)} 0;
-  }
-`
-export const DescriptionTitle = styled(Text)`
-  margin-bottom: ${sizes(2)};
-`
-
-export const DescriptionCopy = styled(Text)`
-  word-break: break-word;
-`
-
-export const Category = styled(Link)`
-  display: flex;
-  align-items: center;
-  text-decoration: none;
-
-  svg {
-    margin-right: ${sizes(2)};
-  }
-`
-
-export const LicenceCategoryWrapper = styled.div<{ detailsExpanded: boolean }>`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-template-rows: 1fr;
-  gap: ${sizes(6)};
-  visibility: ${({ detailsExpanded }) => (detailsExpanded ? 'visible' : 'hidden')};
-  max-height: ${({ detailsExpanded }) => (detailsExpanded ? 'auto' : '0')};
-  overflow: ${({ detailsExpanded }) => (detailsExpanded ? 'unset' : 'hidden')};
-  opacity: ${({ detailsExpanded }) => (detailsExpanded ? 1 : 0)};
-  transition: opacity 150ms ease-out;
-`
-
-export const LicenseCustomText = styled(Text)`
-  margin-top: ${sizes(2)};
-`
-
-export const ExpandButton = styled(Button)`
-  display: block;
-  margin-top: ${sizes(2)};
-`
-
-export const CategoryWrapper = styled.div`
-  text-align: left;
-
-  path {
-    fill: ${cVar('colorText')};
-  }
-`
-
 export const NotFoundVideoContainer = styled.div`
   display: flex;
   align-items: center;
@@ -179,10 +112,6 @@ export const NotFoundVideoContainer = styled.div`
 
 export const Details = styled.div`
   display: flex;
-`
-
-export const StyledLimitedWidthContainer = styled(LimitedWidthContainer)`
-  margin-top: ${sizes(8)};
 `
 
 export const MoreVideosContainer = styled.div`
@@ -207,8 +136,4 @@ export const SeeMoreButton = styled(Button)`
 
 export const StyledCallToActionWrapper = styled(CallToActionWrapper)`
   margin-top: ${sizes(16)};
-`
-
-export const DescriptionLink = styled(Button)`
-  word-break: break-all;
 `
